Fix nextcloud configuration and other bugs
Nextcloud configuration now uses a pod for the application and the
database. The start up configuration for nextcloud has also been
fixed, the systemd files for pods needs to happen in a particular
order. Make now uses bash for all shell processing, fixes a bug
where echo was not displaying correctly. Added remove option to
configurations.